Refactor optimizers into separate files implementing OptimizerPlugin interface
